script_ssm24_eclipse/requirements.txt
networkx
pycryptodome
